+++ app/models/task.rb
@@ -1,6 +1,7 @@
 class Task < ApplicationRecord
   self.implicit_order_column = "created_at"
   belongs_to :section
+  has_many :steps, dependent: :destroy
 
   validates :title, :contentful_id, presence: true
 end